mito-ai: move message history utils into core
