+++ mito-ai/mito_ai/completions/agent_loop.py
@@ -14,10 +14,7 @@ from typing import Any, Callable, List, Optional
 
 from openai.types.chat import ChatCompletionMessageParam
 
-from mito_ai.completions.completion_handlers.utils import (
-    append_agent_system_message,
-    create_ai_optimized_message,
-)
+from mito_ai.completions.completion_handlers.utils import create_ai_optimized_message
 from mito_ai.completions.jupyter_lab_tool_executor import JupyterLabToolExecutor
 from mito_ai_core.completions.message_history import GlobalMessageHistory
 from mito_ai.completions.models import (
