+++ mito-ai/mito_ai/completions/completion_handlers/utils.py
@@ -11,9 +11,6 @@ from openai.types.chat import ChatCompletionMessageParam
 from mito_ai_core.completions.prompt_builders.chat_system_message import (
     create_chat_system_message_prompt,
 )
-from mito_ai_core.completions.prompt_builders.agent_system_message import (
-    create_agent_system_message_prompt,
-)
 
 
 async def append_chat_system_message(
@@ -30,36 +27,6 @@ async def append_chat_system_message(
         return
 
     system_message_prompt = create_chat_system_message_prompt()
-
-    system_message: ChatCompletionMessageParam = {
-        "role": "system",
-        "content": system_message_prompt,
-    }
-
-    await message_history.append_message(
-        ai_optimized_message=system_message,
-        display_message=system_message,
-        llm_provider=provider,
-        thread_id=thread_id,
-    )
-
-
-async def append_agent_system_message(
-    message_history: GlobalMessageHistory,
-    provider: ProviderManager,
-    thread_id: ThreadID,
-    is_chrome_browser: bool,
-) -> None:
-
-    # If the system message already exists, do nothing
-    if any(
-        msg["role"] == "system"
-        for msg in message_history.get_ai_optimized_history(thread_id)
-    ):
-        return
-
-    include_cell_output_tool = is_chrome_browser
-    system_message_prompt = create_agent_system_message_prompt(include_cell_output_tool)
 
     system_message: ChatCompletionMessageParam = {
         "role": "system",
